Moved backup helpers to suite_test_data.go
